Define data-driven tests for parsing and evaluation
`ParserTest` is now dedicated to parsing, by validating that a given input results in a specific AST. The new
`EvaluationTest` is the analogue for checking that the correct answer is received after evaluating a given input,
regardless of the internal structure.

The bulk of the test cases are now captured in text files and are run through a parameterized tests. This avoids
checking multiple cases in a single execution, and avoids the need for so many test methods that essentially do that
same thing.
